# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateThreadRequest": {
        "additionalProperties": false,
        "properties": {
          "goal": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Goal",
            "type": "string"
          }
        },
        "required": [
          "goal"
        ],
        "title": "CreateThreadRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ProductCandidateView": {
        "additionalProperties": false,
        "properties": {
          "evidence_refs": {
            "items": {
              "type": "string"
            },
            "title": "Evidence Refs",
            "type": "array"
          },
          "fit_summary": {
            "default": "",
            "title": "Fit Summary",
            "type": "string"
          },
          "matched_constraints": {
            "items": {
              "type": "string"
            },
            "title": "Matched Constraints",
            "type": "array"
          },
          "product_id": {
            "format": "uuid",
            "title": "Product Id",
            "type": "string"
          },
          "risks": {
            "items": {
              "type": "string"
            },
            "title": "Risks",
            "type": "array"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "unmet_constraints": {
            "items": {
              "type": "string"
            },
            "title": "Unmet Constraints",
            "type": "array"
          },
          "variant_id": {
            "format": "uuid",
            "title": "Variant Id",
            "type": "string"
          }
        },
        "required": [
          "product_id",
          "variant_id",
          "title"
        ],
        "title": "ProductCandidateView",
        "type": "object"
      },
...
      "ThreadCreated": {
        "additionalProperties": false,
        "properties": {
          "mission_id": {
            "format": "uuid",
            "title": "Mission Id",
            "type": "string"
          },
          "thread_id": {
            "format": "uuid",
            "title": "Thread Id",
            "type": "string"
          }
        },
        "required": [
          "thread_id",
          "mission_id"
        ],
        "title": "ThreadCreated",
        "type": "object"
      },
      "ThreadSnapshot": {
        "additionalProperties": false,
        "properties": {
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/ProductCandidateView"
            },
            "title": "Candidates",
            "type": "array"
          },
          "goal": {
            "title": "Goal",
            "type": "string"
          },
          "last_event_id": {
            "minimum": 0.0,
            "title": "Last Event Id",
            "type": "integer"
          },
          "mission_id": {
            "format": "uuid",
            "title": "Mission Id",
            "type": "string"
          },
          "pending_action": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pending Action"
          },
          "status": {
            "enum": [
              "IDLE",
              "RUNNING",
              "WAITING_APPROVAL",
              "WAITING_CLARIFICATION",
              "COMPLETED",
              "FAILED"
            ],
            "title": "Status",
            "type": "string"
          },
          "thread_id": {
            "format": "uuid",
            "title": "Thread Id",
            "type": "string"
          }
        },
        "required": [
          "thread_id",
          "mission_id",
          "goal",
          "status",
          "last_event_id"
        ],
        "title": "ThreadSnapshot",
        "type": "object"
      },
...
      "TurnAccepted": {
        "additionalProperties": false,
        "properties": {
          "event_count": {
            "title": "Event Count",
            "type": "integer"
          },
          "replayed": {
            "title": "Replayed",
            "type": "boolean"
          },
          "run_id": {
            "format": "uuid",
            "title": "Run Id",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "replayed",
          "event_count"
        ],
        "title": "TurnAccepted",
        "type": "object"
      },
      "TurnRequest": {
        "additionalProperties": false,
        "properties": {
          "media_ids": {
            "default": [],
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "maxItems": 8,
            "title": "Media Ids",
            "type": "array"
          },
          "text": {
            "default": "",
            "maxLength": 10000,
            "title": "Text",
            "type": "string"
          }
        },
        "title": "TurnRequest",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Unified multimodal shopping turns, bounded media references and resumable public Agent events.",
    "title": "RAG Commerce Shopping Agent API",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/agent-runs/{run_id}/events": {
      "get": {
        "operationId": "getAgentRunEvents",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Last-Event-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Last-Event-Id"
            }
          },
          {
            "in": "header",
            "name": "x-user-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Events",
        "tags": [
          "agent"
        ]
      }
    },
...
    "/v1/threads": {
      "post": {
        "operationId": "createThread",
        "parameters": [
          {
            "in": "header",
            "name": "x-user-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateThreadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThreadCreated"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Thread",
        "tags": [
          "threads"
        ]
      }
    },
    "/v1/threads/{thread_id}": {
      "get": {
        "operationId": "getThread",
        "parameters": [
          {
            "in": "path",
            "name": "thread_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Thread Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThreadSnapshot"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Thread",
        "tags": [
          "threads"
        ]
      }
    },
    "/v1/threads/{thread_id}/turns": {
      "post": {
        "operationId": "createTurn",
        "parameters": [
          {
            "in": "path",
            "name": "thread_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Thread Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          },
          {
            "in": "header",
            "name": "x-user-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TurnRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TurnAccepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Turn",
        "tags": [
          "agent"
        ]
      }
    },
...
  }
}